requirements.txt
requests==2.24.0
Django==3.0.8
gunicorn==20.0.4
django-crispy-forms==1.9.1
social-auth-app-django==4.0.0
Pillow==7.2.0
pylint==2.5.3
django-bootstrap4==2.2.0 
djangorestframework==3.11.1
psycopg2-binary==2.8.5
python-dotenv==0.14.0 
#django-environ==0.4.5
django-cities-light==3.6.0
django-grappelli==2.14.2
django-ckeditor
sorl-thumbnail
django-modeltranslation
